+++ src/config/config.ts
@@ -18,7 +18,10 @@ export type OptimizationOptions = {
      * If set not to `"disable"`, stores internal and external receivers outside the methods map. Default is `"explorers-compatible"`.
      * Saves gas, but as a result of this optimization, the contract might not be correctly recognized and parsed by indexers.
      */
-    readonly internalExternalReceiversOutsideMethodsMap?: "disable" | "explorers-compatible" | "fast";
+    readonly internalExternalReceiversOutsideMethodsMap?:
+        | "disable"
+        | "explorers-compatible"
+        | "fast";
 };
 
 export type ExperimentalOptions = {